Move dispose logic to Texture but keep finalizer at RefCount
