+++ osu.Framework/Graphics/Lines/Path.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-using System;
 using osu.Framework.Graphics.Primitives;
 using osu.Framework.Graphics.Textures;
 using OpenTK;
@@ -141,7 +140,7 @@ namespace osu.Framework.Graphics.Lines
 
         protected override void Dispose(bool isDisposing)
         {
-            (texture as IDisposable)?.Dispose();
+            texture.Dispose();
             texture = null;
 
             base.Dispose(isDisposing);
@@ -185,7 +184,7 @@ namespace osu.Framework.Graphics.Lines
                 if (value == texture)
                     return;
 
-                (texture as IDisposable)?.Dispose();
+                texture.Dispose();
 
                 texture = value;
                 Invalidate(Invalidation.DrawNode);
