+++ osu.Framework/Graphics/UserInterface/CircularProgress.cs
@@ -25,8 +25,6 @@ namespace osu.Framework.Graphics.UserInterface
 
         private readonly CircularProgressDrawNodeSharedData pathDrawNodeSharedData = new CircularProgressDrawNodeSharedData();
 
-        public bool CanDisposeTexture { get; protected set; }
-
         #region Disposal
 
         protected override void Dispose(bool isDisposing)
